refactor: Refctor into logical abstractions
