+++ package.json
@@ -1,41 +1,42 @@
 {
-    "name": "github-runner-monitor",
-    "version": "1.0.0",
-    "description": "Monitor GitHub self-hosted runners and alert on unhealthy states",
-    "main": "dist/probe.js",
-    "scripts": {
-        "prebuild": "rm -rf dist",
-        "build": "esbuild src/*.ts --bundle --minify --sourcemap --platform=node --target=es2020 --outdir=dist",
-        "watch": "tsc -w",
-        "cdk": "cdk",
-        "deploy": "npm run build && cdk deploy",
-        "synth": "cdk synth",
-        "lint": "eslint .",
-        "lint:fix": "eslint . --fix",
-        "prettier": "prettier --write '**/*.ts'"
-    },
-    "dependencies": {
-        "@aws-sdk/client-ssm": "^3.777.0",
-        "@stately-cloud/client": "^0.28.0",
-        "@types/node-fetch": "^2.6.12",
-        "@typescript-eslint/eslint-plugin": "^8.29.0",
-        "@typescript-eslint/parser": "^8.29.0",
-        "aws-cdk-lib": "^2.187.0",
-        "axios": "^1.8.4",
-        "constructs": "^10.4.2",
-        "esbuild": "^0.25.2",
-        "prettier": "^3.5.3"
-    },
-    "devDependencies": {
-        "@eslint/js": "^9.24.0",
-        "@types/node": "^18.16.18",
-        "aws-cdk": "^2.84.0",
-        "eslint": "^9.24.0",
-        "source-map-support": "^0.5.21",
-        "ts-node": "^10.9.1",
-        "typescript": "^5.1.3",
-        "typescript-eslint": "^8.29.0"
-    },
-    "author": "",
-    "license": "MIT"
+  "name": "github-runner-monitor",
+  "version": "1.0.0",
+  "description": "Monitor GitHub self-hosted runners and alert on unhealthy states",
+  "main": "dist/index.js",
+  "scripts": {
+    "prebuild": "rm -rf dist",
+    "build": "esbuild src/*.ts --bundle --minify --sourcemap --platform=node --target=es2020 --outdir=dist",
+    "watch": "tsc -w",
+    "cdk": "cdk",
+    "deploy": "npm run build && cdk deploy",
+    "synth": "cdk synth",
+    "lint": "eslint .",
+    "lint:fix": "eslint . --fix",
+    "prettier": "prettier --write '**/*.ts'"
+  },
+  "dependencies": {
+    "@aws-sdk/client-ssm": "^3.777.0",
+    "@stately-cloud/client": "^0.28.0",
+    "@types/node-fetch": "^2.6.12",
+    "@typescript-eslint/eslint-plugin": "^8.29.0",
+    "@typescript-eslint/parser": "^8.29.0",
+    "aws-cdk-lib": "^2.187.0",
+    "axios": "^1.8.4",
+    "constructs": "^10.4.2",
+    "esbuild": "^0.25.2",
+    "prettier": "^3.5.3"
+  },
+  "devDependencies": {
+    "@eslint/js": "^9.24.0",
+    "@types/aws-lambda": "^8.10.149",
+    "@types/node": "^18.16.18",
+    "aws-cdk": "^2.84.0",
+    "eslint": "^9.24.0",
+    "source-map-support": "^0.5.21",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.1.3",
+    "typescript-eslint": "^8.29.0"
+  },
+  "author": "",
+  "license": "MIT"
 }
