+++ template.yaml
@@ -17,7 +17,7 @@ Resources:
           AWS_REGION:
           SLACK_WEBHOOK:
       CodeUri: ./dist/
-      Handler: probe.handler
+      Handler: index.handler
       Runtime: nodejs18.x
       Timeout: 300
       MemorySize: 512
